refactor: 앨범 분석 페이지 스트리밍 섹션 컴포넌트 분리 및 스트리밍 클릭 비율 반올림 처리
