+++ src/app/(default)/album/analysis/_client-album-analysis.tsx
@@ -1,18 +1,16 @@
 "use client";
 
 import { Button } from "@/components/ui/button";
-import { Progress } from "@/components/ui/progress";
 import { Separator } from "@/components/ui/separator";
 import { Spinner } from "@/components/ui/spinner";
 import BackButton from "@/components/common/back-button";
 import AlbumAnalysisActionButton from "@/components/mypage/album-analysis-action-button";
+import StreamingSection from "@/components/mypage/analysis-streaming-section";
 import DiagnosisItemCard from "@/components/mypage/diagnosis-item-card";
 import { LinkIcon, CirclePlayIcon } from "lucide-react";
 import Image from "next/image";
 import { useEffect, useState } from "react";
 import { getAnalysisPage } from "@/lib/api/music-promotion";
-import { getStreamingCode } from "@/utils/album";
-import { mapToStreamingBtn } from "@/utils/mapper";
 import { formatDate } from "@/utils/date";
 import { GetAnalysisPageRes } from "@/types/api-response";
 
@@ -134,47 +132,7 @@ export default function AlbumAnalysisPage({ promotionId }: Props) {
 
         <Separator className="-mx-5" />
 
-        <section className="mb-1 flex flex-col gap-6">
-          <h4 className="h4-bold text-font-basic">
-            어떤 스트리밍 사이트로 이동했나요?
-          </h4>
-
-          <ul className="flex flex-col gap-5">
-            {data.streamingLinks.map((item) => {
-              const streamingCode = getStreamingCode(item.url);
-              const streamingItem = mapToStreamingBtn[streamingCode!];
-
-              return (
-                <li
-                  key={streamingItem.label}
-                  className="flex justify-between px-1 py-3"
-                >
-                  <div className="flex items-center gap-3">
-                    <Image
-                      src={streamingItem.icon}
-                      alt={streamingItem.label}
-                      width={32}
-                      height={32}
-                      className="shrink-0"
-                    />
-
-                    <span className="p1-bold-leading-none text-font-middle">
-                      {streamingItem.label}
-                    </span>
-                  </div>
-
-                  <div className="flex items-center gap-2">
-                    <Progress className="w-20" value={item.clickShareRate} />
-
-                    <span className="p2-semibold text-font-middle relative top-px w-10 text-left">
-                      {item.clickShareRate}%
-                    </span>
-                  </div>
-                </li>
-              );
-            })}
-          </ul>
-        </section>
+        <StreamingSection streamingLinks={data.streamingLinks} />
 
         <Separator className="-mx-5" />
